requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
dj-static==0.0.6
Django==4.0.1
django-bootstrap4==21.2
django-chartjs==2.3.0
django-stdimage==5.3.0
Pillow==9.0.0
psycopg2-binary==2.9.3
soupsieve==2.3.1
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
django-mptt==0.13.4
django-extensions